build: Remove Residual Expat
Google originally introduced EXPAT as a dependency to gdb 7.3.

Now 5 versions of gdb later it seems this dependency is no longer needed (I tested this commit on aarch64, arm-eabi, and androideabi and they all built fine. The toolchains also worked fine on a ROM build I ran afterwards).

Furthermore building expat adds an additional 15 seconds or so to build time which may not seem like a lot but it adds up over time. Also keep in mind that 15 seconds is 3-5% of the build time when compiling arm-eabi toolchains so why waste our time with this? (we have better things to do)

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>